Add unpacked boost.
Myrimatch/pwiz needs Boost, but includes an old version (1.43.0)
that needs to be patched in order to compile with gcc 4.9
(default in Debian jessie/stable).

It is not obvious how to change to build to use the system's
Boost. (Besides, the build links statically, which is
undesirable.)

Add the unpacked Boost sources to patch it.

Add unpacked boost in the same location where it would be unpacked
normally by the build system; placing these files in another
location causes link errors because the build does not use the
correct paths:
* Compiling the following object files with boost in a different
  location adds the non-existing path
  -I/<myrimatch_path>/freicore/pwiz_src/<subdir_of_myrimatch_path>/boost_1_43_0
  to the compile command instead of
  -I/<myrimatch_path>/freicore/libraries/boost_1_43_0
  when the default location is used.
  * Configuration_mz5.o
  * Connection_mz5.o
  * Datastructures_mz5.o
  * ReferenceRead_mz5.o
  * ReferenceWrite_mz5.o
  * Translator_mz5.o
* The .o files are combined into libpwiz_data_msdata_mz5.a with
  ar and ranlib.
* Linking the myrimatch executable fails.